reorg subcmds
