+++ src/dswell/subcmds/__init__.py
@@ -1,0 +1,1 @@
+"""Subcommands for the dswell CLI."""
